requirements.txt
asgiref==3.4.1
attrs==23.1.0
build==0.10.0
CacheControl==0.13.1
calculation==8.0
Calculations==1.10
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cleo==2.0.1
colorama==0.4.6
contourpy==1.1.0
crashtest==0.4.1
crispy-bootstrap5==0.7
cycler==0.11.0
Cython==3.0.4
defusedxml==0.7.1
diff-match-patch==20230430
distlib==0.3.7
Django==3.2.13
django-calculation==1.0.0
django-crispy-forms==2.0
django-formtools==2.4.1
django-grappelli==3.0.8
django-import-export==3.3.1
django-ranged-response==0.2.0
django-recaptcha==3.0.0
docopt==0.6.2
dulwich==0.21.6
et-xmlfile==1.1.0
filelock==3.12.4
fonttools==4.42.1
idna==3.4
importlib-metadata==6.8.0
installer==0.7.0
jaraco.classes==3.3.0
jsonschema==4.17.3
keyring==24.2.0
kiwisolver==1.4.5
MarkupPy==1.14
matplotlib==3.7.3
more-itertools==10.1.0
msgpack==1.0.5
multipage-form @ git+https://github.com/ImaginaryLandscape/django-multipage-form.git@12fde2afb819b9e6641e880a7b491679f8f76b1c
numpy==1.25.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.1
pexpect==4.8.0
Pillow==10.0.0
pipreqs==0.4.13
pkginfo==1.9.6
platformdirs==3.10.0
poetry==1.6.1
poetry-core==1.7.0
poetry-plugin-export==1.5.0
ptyprocess==0.7.0
pycparser==2.21
pyparsing==3.1.1
pyproject_hooks==1.0.0
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32-ctypes==0.2.2
PyYAML==6.0.1
rapidfuzz==2.15.1
requests==2.31.0
requests-toolbelt==1.0.0
shellingham==1.5.3
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.1
trove-classifiers==2023.8.7
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
virtualenv==20.24.5
xlrd==2.0.1
xlwt==1.3.0
yarg==0.1.9
zipp==3.16.2
